requirements.txt
wheel>=0.42.0
numpy>=1.26.3
setuptools>=69.0.3
Flask==2.0.1
google-auth-oauthlib==0.4.6
google-auth==2.3.3
earthengine-api==0.1.317
pandas==2.2.3
requests==2.27.1
Werkzeug==2.0.3
python-dotenv==0.19.2
oauth2client==4.1.3
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
